Refactoring - remove 'vm' static functions
